Axios Race Condition Fix
This commit improves the authCheck middleware and related functions responsible for authentication and session management in the application. This commit introduces a new approach for handling CSRF tokens to address a peculiar issue encountered with Axios interceptors. Previously, the CSRF token was tied to the HTTPOnly session cookie, but this caused conflicts with cookies received from subsequent requests before the Axios interceptor set the header, leading to authentication rejections (i.e: sending a request with a CSRF cookie but then the interceptor reads the cookies from another library and sets the header to a cookie that was *just* received (race condition). This makes it so that the CSRF cookie and header don't match. We no longer care about them matching, as long as there is a csrf table entry with their client id existing in the db. This is slightly similar to the 'double-mount' race-condition error that I ran into at the beginning of my work on this project. To address this, a new table-based approach is implemented, decoupling CSRF tokens from user sessions. Similar to how login sessions are deleted 15s after validation, the same applies to CSRF sessions too. This means that the user can validate using any of the CSRF tokens given to them recently (or in the case of the axios race condition, they can use a csrf cookie header from a cookie that was received while axios was building a request with a cookie from before)